Change option name from batched_pack_unpack to batched_copy, move NCCL related class/method to pure_nccl_communicator.py, etc.
