+++ chainermn/communicators/__init__.py
@@ -3,7 +3,7 @@ from chainermn.communicators.communicator_base import CommunicatorBase  # NOQA
 
 def create_communicator(
         communicator_name='hierarchical', mpi_comm=None,
-        allreduce_grad_dtype=None, batched_pack_unpack=False):
+        allreduce_grad_dtype=None, batched_copy=False):
     """Create a ChainerMN communicator.
 
     Different communicators provide different approaches of communication, so
@@ -59,9 +59,9 @@ def create_communicator(
         raise ValueError(
             'allreduce_grad_dtype is only available'
             'at \'pure_nccl\' communicator.')
-    if communicator_name != 'pure_nccl' and batched_pack_unpack:
+    if communicator_name != 'pure_nccl' and batched_copy:
         raise ValueError(
-            'batched_pack_unpack is only available'
+            'batched_copy is only available'
             'at \'pure_nccl\' communicator.')
 
     if communicator_name == 'naive':
@@ -99,7 +99,7 @@ def create_communicator(
             import PureNcclCommunicator
         return PureNcclCommunicator(mpi_comm=mpi_comm,
                                     allreduce_grad_dtype=allreduce_grad_dtype,
-                                    batched_pack_unpack=batched_pack_unpack)
+                                    batched_copy=batched_copy)
 
     elif communicator_name == 'dummy':
         from chainermn.communicators.dummy_communicator \
