+++ chainermn/communicators/_memory_utility.py
@@ -112,170 +112,6 @@ def unpack_params(params, itemsize, attr_name, buffer, stream=None):
         offset += size
 
 
-class ParamsData(object):
-    def __init__(self, params, attr_name):
-        n_params = len(params)
-        params_dptr = np.empty(n_params, dtype=np.int64)
-        params_dtype = np.empty(n_params, dtype=np.int32)
-        params_size_csum = np.empty(n_params+1, dtype=np.int32)
-        params_size_csum[0] = 0
-        for i, param in enumerate(params):
-            v = getattr(param, attr_name)
-            params_dptr[i] = v.data.ptr
-            params_dtype[i] = _get_nccl_type_id(v.dtype)
-            params_size_csum[i+1] = params_size_csum[i] + v.size
-        self.n_params = n_params
-        self.n_elems = params_size_csum[n_params]
-        self.params_size_csum = cp.asarray(params_size_csum)
-        self.params_dtype = cp.asarray(params_dtype)
-        self.params_dptr = cp.asarray(params_dptr)
-
-
-def batched_pack_params(params_data, buffer, dtype):
-    n_params = params_data.n_params
-    n_elems = params_data.n_elems
-    params_dptr = params_data.params_dptr
-    params_dtype = params_data.params_dtype
-    params_size_csum = params_data.params_size_csum
-    buf_dtype = _get_nccl_type_id(dtype)
-    n_threads = 128
-    n_blocks = (n_elems + n_threads - 1) // n_threads
-    _cupy_batched_pack_params()(
-        (n_blocks, ), (n_threads, ),
-        (buffer.memory.ptr, buf_dtype, n_elems,
-         params_dptr, params_dtype, params_size_csum, n_params))
-
-
-def batched_unpack_params(params_data, buffer, dtype):
-    n_params = params_data.n_params
-    n_elems = params_data.n_elems
-    params_dptr = params_data.params_dptr
-    params_dtype = params_data.params_dtype
-    params_size_csum = params_data.params_size_csum
-    buf_dtype = _get_nccl_type_id(dtype)
-    n_threads = 128
-    n_blocks = (n_elems + n_threads - 1) // n_threads
-    _cupy_batched_unpack_params()(
-        (n_blocks, ), (n_threads, ),
-        (buffer.memory.ptr, buf_dtype, n_elems,
-         params_dptr, params_dtype, params_size_csum, n_params))
-
-
-def _cupy_batched_pack_params():
-    return cp.RawKernel(r'''
-#include <cuda_fp16.h>
-#define NCCL_FLOAT16  6
-#define NCCL_FLOAT32  7
-    extern "C" __global__
-    void cupy_batched_pack_params(
-            void *dst0, int dst_dtype, int n_elems,
-            unsigned long *params_dptr, int *params_dtype,
-            int *params_size_csum, int n_params) {
-        int tid = threadIdx.x + blockIdx.x * blockDim.x;
-        if (tid >= n_elems) return;
-        int j_min = 0;
-        int j_max = n_params - 1;
-        int j;
-        while (1) {
-            j = (j_min + j_max) / 2;
-            if (tid < params_size_csum[j]) {
-                j_max = j - 1;
-                continue;
-            }
-            if (tid >= params_size_csum[j+1]){
-                j_min = j + 1;
-                continue;
-            }
-            break;
-        }
-        assert(tid >= params_size_csum[j]);
-        assert(tid < params_size_csum[j+1]);
-        int src_dtype = params_dtype[j];
-        int src_idx = tid - params_size_csum[j];
-        if (dst_dtype == NCCL_FLOAT16) {
-            half* dst = (half*) dst0;
-            if (src_dtype == NCCL_FLOAT16) {
-                dst[tid] = (half) (((half*) (params_dptr[j]))[src_idx]);
-            }
-            else if (src_dtype == NCCL_FLOAT32) {
-                dst[tid] = (half) (((float*) (params_dptr[j]))[src_idx]);
-            }
-        }
-        else if (dst_dtype == NCCL_FLOAT32) {
-            float* dst = (float*) dst0;
-            if (src_dtype == NCCL_FLOAT16) {
-                dst[tid] = (float) (((half*) (params_dptr[j]))[src_idx]);
-            }
-            else if (src_dtype == NCCL_FLOAT32) {
-                dst[tid] = (float) (((float*) (params_dptr[j]))[src_idx]);
-            }
-       }
-    }
-    ''', 'cupy_batched_pack_params')
-
-
-def _cupy_batched_unpack_params():
-    return cp.RawKernel(r'''
-#include <cuda_fp16.h>
-#define NCCL_FLOAT16  6
-#define NCCL_FLOAT32  7
-    extern "C" __global__
-    void cupy_batched_unpack_params(
-            void *src0, int src_dtype, int n_elems,
-            unsigned long *params_dptr, int *params_dtype,
-            int *params_size_csum, int n_params) {
-        int tid = threadIdx.x + blockIdx.x * blockDim.x;
-        if (tid >= n_elems) return;
-        int j_min = 0;
-        int j_max = n_params - 1;
-        int j;
-        while (1) {
-            j = (j_min + j_max) / 2;
-            if (tid < params_size_csum[j]) {
-                j_max = j - 1;
-                continue;
-            }
-            if (tid >= params_size_csum[j+1]){
-                j_min = j + 1;
-                continue;
-            }
-            break;
-        }
-        assert(tid >= params_size_csum[j]);
-        assert(tid < params_size_csum[j+1]);
-        int dst_dtype = params_dtype[j];
-        int dst_idx = tid - params_size_csum[j];
-        if (src_dtype == NCCL_FLOAT16) {
-            half* src = (half*) src0;
-            if (dst_dtype == NCCL_FLOAT16) {
-                ((half*) (params_dptr[j]))[dst_idx] = (half) src[tid];
-            }
-            else if (src_dtype == NCCL_FLOAT32) {
-                ((float*) (params_dptr[j]))[dst_idx] = (float) src[tid];
-            }
-        }
-        else if (src_dtype == NCCL_FLOAT32) {
-            float* src = (float*) src0;
-            if (dst_dtype == NCCL_FLOAT16) {
-                ((half*) (params_dptr[j]))[dst_idx] = (half) src[tid];
-            }
-            else if (src_dtype == NCCL_FLOAT32) {
-                ((float*) (params_dptr[j]))[dst_idx] = (float) src[tid];
-            }
-       }
-    }''', 'cupy_batched_unpack_params')
-
-
-def _get_nccl_type_id(dtype):
-    if dtype == np.float16:
-        return 6  # nccl.NCCL_FLOAT16
-    elif dtype == np.float32:
-        return 7  # nccl.NCCL_FLOAT32
-    else:
-        raise ValueError(
-            'dtype must be float16 or float32.')
-
-
 def array_to_buffer_object(array, mpi_dtype=mpi4py.MPI.FLOAT):
     xp = chainer.backend.get_array_module(array)
